Solution to Input and Print in Hexadecimal and Octal

+++ Input and Print in Hexadecimal and Octal@@6788d5012db25efec946f941/code.c
@@ -1,6 +1,6 @@
 #include <stdio.h>
 int main() {
-    float a;
+    int a;
     printf("");
     scanf("%d",&a);
     printf("Hexadecimal: %x\n",a);
